// freelancer-client/src/features/sellers/interfaces/seller.interface.ts
export interface ICertificate {
  [key: string]: string | number | undefined;
  _id?: string;
  name: string;
  from: string;
  year: number | string;
}
// ...
export interface ICertificateProps {
  selectedField?: ICertificate;
  certificatesFields?: ICertificate[];
  setCertificatesFields?: Dispatch<SetStateAction<ICertificate[]>>;
  setShowCertificateAddForm?: Dispatch<SetStateAction<boolean>>;
  setShowCertificateEditForm?: Dispatch<SetStateAction<boolean>>;
}